Add OperationalContext as its own component
