+++ src/OperationalUI/README.md
@@ -1,4 +1,4 @@
-Main provider for Operational UI. Should need to wrap all your application with this component.
+Main wrapper for Operational UI, required at the top level of the application. See `OperationalContext` for some more involved context-reliant features of this component.
 
 ### Classic example
 
@@ -56,92 +56,4 @@ class RoutingComponent extends React.Component {
 }
 
 ;<RoutingComponent />
-```
-
-### Window size
-
-This component keeps track of the window size and makes it available through context, as in the snippet below:
-
-```jsx
-// This import statement is necessary because of the wiring of this demo website.
-// When using the library outside, use the following:
-// `import { OperationalUI, OperationalContext } from "@operational/components"`
-const OperationalContext = require("../").OperationalContext
-;<OperationalUI>
-  <OperationalContext>
-    {operationalContext => (
-      <p>{`The viewport is ${operationalContext.windowSize.width} pixels wide and ${
-        operationalContext.windowSize.height
-      } tall.`}</p>
-    )}
-  </OperationalContext>
-</OperationalUI>
-```
-
-### Messages and Loaders
-
-You can use `OperationalUI`'s flash- and progress bar features to automatically render and manage these universal UI elements using the `pushMessage` and `setLoadingState` methods provided in context, as shown in the code snippet below:
-
-```jsx
-// This import statement is necessary because of the wiring of this demo website.
-// When using the library outside, use the following:
-// `import { OperationalUI, OperationalContext, Button } from "@operational/components"`
-const OperationalContext = require("../").OperationalContext
-
-class MessageHandler extends React.Component {
-  render() {
-    return (
-      <OperationalUI>
-        <OperationalContext>
-          {operationalContext => (
-            <div style={{ padding: 20 }}>
-              <Button
-                color="primary"
-                onClick={() => {
-                  operationalContext.pushMessage({
-                    body: "Info message",
-                    type: "info",
-                  })
-                }}
-              >
-                Create an info message
-              </Button>
-              <Button
-                color="success"
-                onClick={() => {
-                  operationalContext.pushMessage({
-                    body: "Success message",
-                    type: "success",
-                  })
-                }}
-              >
-                Create a success message
-              </Button>
-              <Button
-                color="error"
-                onClick={() => {
-                  operationalContext.pushMessage({
-                    body: "Error message",
-                    type: "error",
-                  })
-                }}
-              >
-                Create an error message
-              </Button>
-              <Button
-                onClick={() => {
-                  operationalContext.setLoading(!operationalContext.loading)
-                }}
-              >
-                Toggle loading state
-              </Button>
-            </div>
-          )}
-        </OperationalContext>
-      </OperationalUI>
-    )
-  }
-}
-
-;<MessageHandler />
 ```
