+++ src/index.ts
@@ -30,7 +30,8 @@ export { default as Input } from "./Input/Input"
 export { default as Message } from "./Message/Message"
 export { default as Messages } from "./Messages/Messages"
 export { default as NameTag } from "./NameTag/NameTag"
-export { default as OperationalUI, Context, Consumer as OperationalContext } from "./OperationalUI/OperationalUI"
+export { default as OperationalUI } from "./OperationalUI/OperationalUI"
+export { default as OperationalContext, Context } from "./OperationalContext/OperationalContext"
 export { default as Page } from "./Page/Page"
 export { default as PageArea } from "./PageArea/PageArea"
 export { default as PageAreas } from "./PageAreas/PageAreas"
